[breaking] Remove automatic unsubscription from the Subscription object.
Summary:
This means that the disconnect method of observables will no longer be automatically invoked when the Subscription object is released.

This has several advantages:

- Subscriptions no longer need to be explicitly maintained if unsubscription is unimportant to the subscriber.
- The subscribe method's return value can now be marked ignorable, simplifying any call sites that don't care about unsubscribing.

The documentation for IndefiniteObservable has been updated to reflect the new behavior. Notably, we'd previously written that Subscriptions maintained a reference to the IndefiniteObservable chain; this isn't/wasn't technically true. Subscriptions only hold references to the chain of disconnect methods.

It is now possible to write expressions like so:

    observable.subscribe { value in
      print(value)
    }

With no explicit subscription storage required.

Reviewers: O2 Material Motion, O4 Material Apple platform reviewers, #material_motion, markwei

Reviewed By: O2 Material Motion, O4 Material Apple platform reviewers, #material_motion, markwei

Tags: #material_motion

Differential Revision: http://codereview.cc/D3116